+++ src/ConfigSymbols/ReceiveLogFunction.d.ts
@@ -1,0 +1,3 @@
+declare const ReceiveLogFunction: unique symbol
+
+export = ReceiveLogFunction