+++ src/ConfigSymbols/ReceiveLogFunction.lua
@@ -1,2 +1,2 @@
 local Symbol = require(script.Parent.Parent.Symbol)
-return Symbol.named("ReceiveLogsFunction")
+return Symbol.named("ReceiveLogFunction")