+++ src/ConfigSymbols/ReceiveLogsFunction.d.ts
@@ -1,3 +1,0 @@
-declare const ReceiveLogsFunction: unique symbol
-
-export = ReceiveLogsFunction