+++ src/ConfigSymbols/SendLogFunction.d.ts
@@ -1,0 +1,3 @@
+declare const SendLogFunction: unique symbol
+
+export = SendLogFunction