+++ src/index.d.ts
@@ -3,7 +3,8 @@ import type ServerBridge from './ServerBridge';
 import type { default as DefaultReceive } from './ConfigSymbols/DefaultReceive';
 import type { default as DefaultSend } from './ConfigSymbols/DefaultSend';
 import type { default as PrintRemotes } from './ConfigSymbols/PrintRemotes';
-import type { default as ReceiveLogsFunction } from './ConfigSymbols/ReceiveLogsFunction';
+import type { default as ReceiveLogFunction } from './ConfigSymbols/ReceiveLogFunction';
+import type { default as SendLogFunction } from './ConfigSymbols/SendLogFunction';
 import type { default as serdeLayer } from './serdeLayer';
 import type { default as rateManager } from './rateManager';
 
@@ -11,7 +12,8 @@ type StartOptions = {
 	[DefaultReceive]: number | null;
 	[DefaultSend]: number | null;
 	[PrintRemotes]: boolean | null;
-	[ReceiveLogsFunction]: (...args: Array<unknown>) => undefined | null;
+	[ReceiveLogFunction]: (...args: Array<unknown>) => undefined | null;
+	[SendLogFunction]: (...args: Array<unknown>) => undefined | null;
 };
 
 type BridgeDictionary = {
@@ -23,6 +25,11 @@ type StringDictionary = {
 }
 
 interface BridgeNet {
+	DefaultReceive: typeof DefaultReceive;
+	DefaultSend: typeof DefaultSend;
+	SendLogFunction: typeof SendLogFunction;
+	ReceiveLogFunction: typeof ReceiveLogFunction;
+	
 	GetReceiveRate: rateManager.GetReceiveRate;
 	SetReceiveRate: rateManager.SetReceiveRate;
 	GetSendRate: rateManager.GetSendRate;